move flow setting logic into its own helper abstraction, deprecate useReducer for FlowRouter
